requirements.txt
PyYAML==6.0.2
python-dateutil==2.9.0.post0